op-cancel-appointment - move bark-appointment and bark-appointment-ids into their own model files
