+++ src/lib/bark/models/bark-appointment-ids.ts
@@ -1,0 +1,10 @@
+import { z } from "zod";
+
+
+export const BarkAppointmentIdsSchema = z.object({
+  appointmentId: z.string(),
+  vetId: z.string(),
+  dogId: z.string(),
+});
+
+export type BarkAppointmentIds = z.infer<typeof BarkAppointmentIdsSchema>;
