+++ src/lib/bark/queries/select-appointment-ids.ts
@@ -1,5 +1,5 @@
 import { DbContext, dbQuery } from "@/lib/data/db-utils";
-import { BarkAppointmentIds, BarkAppointmentIdsSchema } from "../bark-models";
+import { BarkAppointmentIds, BarkAppointmentIdsSchema } from "../models/bark-appointment-ids";
 
 export async function selectAppointmentIds(
   dbContext: DbContext,
